requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=9.0.0
pandas>=1.5.0
plotly>=5.15.0
numpy>=1.21.0
torch>=1.13.0
torchvision>=0.14.0